Fix all existing live service tests - fix paths, add required missing login, fix imports, etc.
